Refactor scripts to use rerank_tools module
Moved core logic from scripts/rerank_ab_test.py, rerank_eval.py, rerank_events.py, rerank_local.py, and rerank_query.py into the new scripts/rerank_tools package for better modularity and maintainability. The original scripts now act as thin shims that import and expose the new implementations, ensuring backward compatibility.
